S2/CircusTrainConsole: big code cleanup + S2/WebApplicationTest: added API project
CircusTrainConsole updates:
changed and relocated AnimalMayBeAdded now know as TryEatOtherAnimal from the wagon class to the Animal class. A animal knows if he can eat another animal, the wagon doesn't know that, the wagon just want's to be filled.

Replaced WagonIsFull function for AnimalIsNotToLarge function.

Code for PutEveryAnimalInAWagon  split up to PutAllCarnivorsInWagon and PutAllHerbivoresInWagon. (still in the same function)

Deleted RemoveEmptyWagons() function. There shouldn't be empty wagons if the code is right.